requirements.txt
fastapi
uvicorn[standard]
pandas
python-multipart
pika
sqlalchemy
psycopg2-binary